Move RealTorControlProcessor to internal file
