+++ library/controller/kmp-tor-controller/src/commonMain/kotlin/io/matthewnelson/kmp/tor/controller/internal/controller/-RealTorController.kt
@@ -78,7 +78,7 @@ internal class RealTorController(
     }
 
     private val processorDelegate: TorControlProcessor by lazy {
-        TorControlProcessor.newInstance(
+        RealTorControlProcessor(
             controlPortInteractor,
             TorControlProcessorLock(
                 dispatcher = dispatcher,
